Map exception types from proxy back to thrift exceptions
Summary:
Because we normally don't serialize TTransportException or TProtocolException,
the proxy serializes these errors as TApplicationException. This makes it more
difficult for a proxied client to distinguish the type of error that actually
happened.

This change makes the SR proxy send "ex" codes in the thrift header of the
response that indicate when the failure was the result of transport or protocol
exceptions, and adds mapping to the clients created by the swift proxy client
utility so that these exceptions will be received properly.

Test Plan: - test this on MQTT, see if it helps alleviate failures from using sr proxy

Reviewed By: alandau@fb.com

FB internal diff: D1167689

@override-unit-failures