thrift: use thrift-py.lpar over thrift-py.par
Summary:
We're seeing startup race-conditions on the build machines when
thrift-py.par is started up by several jobs simultaneously.  This
diff switches fbcode builds to favor the LPAR when it exists to
avoid this issue, which has the additional benefit of improving the
stratup time.

Test Plan:
Ran builds w/ and w/o the LPAR available.  Verified via strace that
the LPAR was used when it was available, and the PAR otherwise (which
is the case when thrift is run from its contbuild publish location
which only has access to the PAR).

Reviewed By: njormrod@fb.com

FB internal diff: D1169592