Addition hook to know when the server is ready
Summary:
TDirectServer currently does not have a hook to notify other
threads that it is ready to serve traffic. This diff implements this
extra hook so that other services that use TDirectServer can have a
means to be notified that it is ready to serve traffic.

Test Plan:
Played with the test servers and verified that it is working
normally.

Reviewed By: andrewcox@fb.com

FB internal diff: D1165692