support thrift functions with more than 1 argument
Summary:
Fix the D thrift code so it compiles when using thrift functions with
multiple arguments.  The code was missing commas in many places when it
tried to join the arguments together.

The error messages from the D compiler seem to be pretty unhelpful in
this case, too.  These errors occur in mixin strings, and the line
numbers in the compile error messages appear to be completely irrelevant
for tracking down where the problem exists.

onelessclown

Test Plan:
notimetoexplaingetintotheclowncar

Updated our tutorial code to show creating a client for an interface
that has a method with 2 arguments.

Reviewed By: ghwang@fb.com

FB internal diff: D1181377