Refactor ThreadManager to use std::chrono, prepare for observable ThreadManager interface
Summary: I plan to include an observer pattern on the ThreadManager similar to TServerObserver. For each item that is queued, I want to know the point in time that it is queued, the point in time it starts work, and the point in time it finishes work. I plan to include notifying the observer in reportStats(), but reportStats() takes in durations instead of points in time. While refactoring this, I figured that switching to more explicit units of time (std::chrono) is a good idea.

Test Plan: reran unit tests

Reviewed By: davejwatson@fb.com

FB internal diff: D1194700