Don't spawn cleanup thread manager if there is nothing to clean
Summary:
If we're not even using security in the client, there is no sense spawning a large thread manager during destruction.

This doesn't directly fix the problem introduced by D1207456, but makes sure that the codepath is never hit. Still need to figure out how to control the destruction order of globals... not sure if there is a good way.

Test Plan: loadgen

Reviewed By: philipp@fb.com

FB internal diff: D1211073