Remove race condition in thread startup in TEventJobQueue
Summary:

In TEventJobQueue::run(), we were setting ##thread_## to the current
thread, and in TEventJobQueue::join(), we are using ##thread_##.  So
code that looks like this has a terrible race condition:

auto thread = std::shared_ptr<JobThread>(new JobThread(this));
threadFactory->newThread(thread)->start();
thread->join();

I changed it so that inside ##newThread()##, when the factory sets the
Thread of its runnable, the JobThread grabs its copy then.

This is only dangerous in the case that ##start()## fails and throws
an exception, but in that case, TEventJobQueue::shutdown() will crash
anyway because some of the elements of ##threads_## will be null.

If we don't want to intercept Runnable::thread(), then we will have to
change JobThread::run() to notify its parent after it sets thread_,
and have TEventJobQueue::init() wait for all its threads to have fully
started before returning.

Test Plan:

I added a test that failed accessing a null about one out of three
times.  It now never fails.

Reviewed By: davejwatson@fb.com

FB internal diff: D1228774

Blame Rev: