Change default client from host/fqdn to first entry in keytab.
Summary:
Client principal choice: first of explicitly specified,
principal in current ccache, or first principal in keytab.  It's
possible for none to be available at startup.  In that case, we just
keep trying until a ccache or keytab appears.  Once a client principal
is chosen, we keep using that one and it never changes.

Test Plan: run regression tests

Reviewed By: andrei.bajenov@fb.com

FB internal diff: D1229336