repro for union list TCompact deserialization bug
Summary:
Hey thrift team, I surfaced this bug where unions getting deserialized out of a list are set to empty.

This is a repro of the bug. Might not be minimal.

Somehow the deserializer extracts a T_STOP TType when reading the union type field, resulting in __clear being
called and an empty union returned. I believe this only happens on the second element or further.

@tgeorgiou looked at the java deserialization code and noticed the difference of an extra readFieldBegin call.
I added that call in a similar spot (included in this patch). It eats the T_STOP and fixes the problem. Disclaimer: I have no idea
whether this fix is appropriate or if it breaks other cases, but I figured it might be a helpful lead so I included it.

Remove the fix to see the breakage.

Test Plan: fbmake runtests

Reviewed By: davejwatson@fb.com

FB internal diff: D1217806