support using fbstring for binary fields in cpp1 code
Summary:
Update the cpp code to allow using fbstring for binary types, just like
we do for string types.  The cpp2 code already supports fbstring for both
binary and string types.

This code is pretty inefficient--it uses the underlying std::string
implementations and then copies to/from a folly::fbstring.  This is
exaclty what the code is already doing for readString()/writeString()
though.

Test Plan:
Verified that a cpp1 client was able to use a thrift file that contains
an fbstring binary type.

Reviewed By: davejwatson@fb.com

FB internal diff: D1247787