Generate TEnumTraits<>::{min,max} for cpp2
Summary:
This diff adds support for the min() and max() functions in TEnumTraits
for cpp2. This support was present in the cpp generator, but cpp2 had a
TODO in the code.

The unit tests are cloned from the tests for the cpp generator and modified
to work with the strongly typed enums generated by cpp2.

Test Plan: fbconfig thrift/test && fbmake runtests

Reviewed By: alandau@fb.com

FB internal diff: D1256488