Add shortEnumNameSafe
Summary:
The current version of `shortEnumName` returns nullptr if the enum value is not found.
This caused SEV 1: https://our.cstools.facebook.com/intern/sevmanager/sev/386289464844573/
Though it supports returning custom name to avoid this, it could be hard to force using of extended version.
That is why we need to create version safe by default.
It will be useful for logging and monitoring purposes to return the actual integer value converted to string.
I will codemod usage of `shortEnumName` in admarket in the next diff.

Test Plan: Run unit tests

Reviewed By: alandau@fb.com

FB internal diff: D1262717