Move TAsyncSSLSocketTest to gtest, update certs
Summary:
This makes it a lot easier to examine the output.

Turns out that the actual issue I was trying to fix was that the CA expired
today.  Replaced all of the crazy generation steps with a bash script that
auto-generates the stuff.

Test Plan: ran the unit tests

Reviewed By: afrind@fb.com

FB internal diff: D1266037