create thrift/lib/cpp/ssl for ssl utilities
Summary:
We're going to need to use this in another place and this is
generally useful code, so let's put it in thrift/lib/cpp/ssl.  Plus, if
we're lucky, we'll get extra eye-balls on it.

Test Plan:
Added unit tests for testing x509 cert validity.  This required
changing the API, but is probably way net positive.

Reviewed By: afrind@fb.com

FB internal diff: D1271497