WIP: changes for building with GHC 7.8
Summary:
This is almost all the changes we need, I think.  I'd like to make the
tree build with both 7.6 and 7.8 for the time being, until we're happy
that 7.8 is working well enough for us, hence the conditional
compilation.

Test Plan:
Current status: almost everything builds, and haxlsh works.
I have a problem with duplicate binary packages that is preventing the
engine unit tests from working.

Reviewed By: jon.coens@fb.com

FB internal diff: D1265259