Add TAsyncSSLSocket::setServerName()
Summary:
- Add TAsyncSSLSocket::setServerName() method to allow callers to set
the string to be used for the server_name TLS extension. This is
useful for applications that don't have full control over the
construction of TAsyncSSLSocket instances, which was the only
way to set this prior to this change.

Test Plan:
- Unit tests
- Ran experimental/rajatgoel2010:get; verified via tcpdump that we're
sending www.facebook.com in the server_name extension

Reviewed By: subodh@fb.com

FB internal diff: D1294060