Renaming objects whose names are reserved in D
Summary:
When building a client in a new languages for an existing service, it might turn out that some identifiers used in the definition of the service (structs' fields, functions' args) are reserved in this language.

In this particular case, the ServiceRouter service uses the 'version' identifier for some fields/function args. Since the 'version' identifier is reserved in D, the generated code fails to compile.

(This diff is a follow-up to https://phabricator.fb.com/D1284635).

Test Plan: Compiled Rapido CLI (fbconfig rapido/cli && fbmake opt).

Reviewed By: andrei.alexandrescu@fb.com

FB internal diff: D1290643