Thrift1 Http protocol and 404 responses
Summary:
When using THttpAsyncChannel on the client in thrift1, a
reponse of 404 throws an error in THttpClientParser::parseStatusLine,
which is caught in TStreamAsyncChannel::invokeReadDataAvailable which
ultimately calls an error callback. If the error callback checks the
error with TStreamAsyncChannel::errorType, it will fail the DCHECK
because the real transport (socket) is good, and THttpAsyncChannel
doesn't override good().

The real fix would be go add a good() implementation to
THttpAsyncChannel, but this is thrift1 and is being phased out anyway.

Previous attempt at fixing in D1273704.

@override-unit-failures

Test Plan: test that it no longer crashes

Reviewed By: andrewcox@fb.com

FB internal diff: D1297027