add entry points for credential-related operations
Summary:
This includes a new Krb5Credentials class which includes a
krb5_creds struct as a member, and methods to get new creds from a
keytab, get new creds from a tgt in a ccache, and store and retrieve
ccache entries.

This does include the first non-trivial wrappers in the form of the
retrieveCred and getCredentials overloads which take a krb5_principal.
This is by far the common use case and the only way we actually use
the functions in the code base.

Test Plan:
refactored separate perf code to use the new methods, and
verified they worked as expected.

Reviewed By: andrei.bajenov@fb.com

FB internal diff: D1301940