Mutex benchmark
Summary:
A load test of a production service showed that replacing a usage of thrift's Mutex with std::mutex had significant performance improvements. This benchmark is designed to help understand why there was such an improvement to try to rule out load test inconsistencies.

The results show that in the happy case, std::mutex is slightly faster in its constructor and in its lock methods. Contention tests are very similar. However, one interesting test case is one in which we thrash jemalloc by making many large malloc calls from lots of threads at the same time. Constructing a thrift mutex can take upwards of 20us (at least) in the test. The production service that saw the improvement from switching is known to have too much contention in malloc, so this may explain the performance win (every malloc is a place where the thread can be blocked on a lock, get context switched out, and sit in the run queue for some time. This service not only hammers malloc, but has a tendency to run completely out of cpu).

Test Plan: This is a benchmark

Reviewed By: davejwatson@fb.com

FB internal diff: D1306318