Make mutex profiling usable by other locks
Summary:
Mutex profiling should not be tied to a particular mutex
implementation. This refactors it out so that we could wrap a std::mutex,
std::recursive_mutex, etc.

The thrift Mutex class is pretty bad. It uses a shared_ptr which requires a
malloc (before this diff, there were *two* mallocs because we were not using
std::make_shared). Unfortunately, we cannot fix this because it breaks the
ability to copy the Mutex (which is another stupid feature of this Mutex).

The introduction of the new mutex allows us to replace user level code that
wants profiling without incurring the cost of the unnecessary malloc.

Test Plan: Existing unit tests exercise the mutex code

Reviewed By: alandau@fb.com

FB internal diff: D1309478