Support utf8strings in fastbinary
Summary: fastbinary doesn't read this flag so it breaks when using fastbinary with unicode in python2. In python 3 fastbinary will encode it into a bytes but still only returns bytes, not unicode. This diff passes the utf8strings to fastbinary so it can read and write unicode in python2 and 3.

Test Plan: Unit tests pass in python 2&3.

Reviewed By: davejwatson@fb.com

FB internal diff: D1264616