Move generated process method to lib
Summary: As title. This simplifies the generator a little and another benefit is I need to change process for supporting asyncio and with this I can do it in lib, not code generator.
@override-unit-failures

Test Plan: fbconfig thrift/test/py && fbmake runtests. Also tested it works with twisted.

Reviewed By: andrewcox@fb.com

FB internal diff: D1319684