Credentials cache management performance improvements
Summary:
Sorry for the large diff, but wasn't sure how to split it up.

The basic idea here is that kerberos credentials caches are slow. They're basically a linked list internally and are copied around inside the gss layer. If the caches are small, this is not a problem, but once they become larger, the perf hit is quite high. We ran into a problem with this when deploying security to the tupperware scheduler. With large credentials caches, there is basically too much lock conention around accessing and maintaining this one linked list. I suspect this might also be causing some timeouts on webis during pushes, where cold starts fire off a lot of requests, and at that point we see a lot of contention even for ccaches ~100.

To help with this, I rewrote the credentials cache manager to maintain a separate credentials cache for each principal. This way, we can grab the correct cache based on where we want to connect. This way, we have full control around how credentials are managed, and we don't need to worry about the internals of kerberos. Unfortunately, there is still a linked list internally in kerberos that maintains all the open memory caches. Fortunately, contention around this list is lower.

Doing this has a few other benefits:
- It allows us to differentiate betwen when we talk to the KDC and when we read from the ccache.
- It allows us to make sure we don't sent duplicate requests to the KDC

Other than modifying the cc manager, I also added a few helpers to Krb5Util.

Test Plan:
Ran a modified version of D1305281. Before this diff, we were getting timeouts 100% of the time, with this diff, almost 100% success. I've tried with credentials caches around 6.5k in size.

I also tested:
- expiring caches
- file cache management (when it's replaced, renewed, etc.)

Reviewed By: mhorowitz@fb.com

Subscribers: mcduff, marccelani, hitesh, mshneer, rkroll, alandau, bmatheny, srenfro, rodrigos, acampi, wstefancik

FB internal diff: D1319424