Thrift asyncio server
Summary:
Implement a thrift server using asyncio, a.k.a tulip. This is Python 3 only, and it only supports header protocol.

Methods in a handler can be decorated and run in 3 different ways. After talking with @andrewcox, running in executor threads is set to default:
1) undecorated, the method is called in the event loop's executor threads,
2) decorated with @thrift.util.asyncio.run_in_loop, the method is called in the event loop thread directly,
3) decorated with @asyncio.coroutine, the method is yielded from in the event loop thread, thus yield from can be used to "chain" other coroutines.

yield from should NOT be used in 1 or 2 because those methods are called directly. yield from turns a method into a coroutine function,
just like yield turns a method into a generator function. Calling a coroutine function does not start it running - it just returns a coroutine.
If yield from is used in 1 or 2 the server will break because when it expects some result it gets a coroutine.

Client isn't implemented yet, send it out to gather feedback first.

Test Plan:
The server in thrift/tutorial/py.asyncio/ uses all 3 kinds of methods. Start it with:
_build/dbg/thrift/tutorial/py.asyncio/py-tutorial-asyncio-server.par

Tested with pyremote client:
_build/dbg/thrift/tutorial/tutorial-asyncio-Calculator-pyremote.par -h 127.0.0.1:8848 ping
_build/dbg/thrift/tutorial/tutorial-asyncio-Calculator-pyremote.par -h 127.0.0.1:8848 calculate 1 "Work(num1=2, num2=3, op=1)"
_build/dbg/thrift/tutorial/tutorial-asyncio-Calculator-pyremote.par -h 127.0.0.1:8848 add 2 3

Once the asyncio client is ready more interesting cases can be tested.

Reviewed By: fried@fb.com

Subscribers: simpkins, ruibalp, agallagher, alandau, bmatheny, andrewcox

FB internal diff: D1324195