un-deprecate `from` as an identifier
Summary:
Using `from` as an identifier was deprecated, with the warning
that it "will soon become a reserved keyword". That was 2011. Since we
still and probably will continue to use `from`, un-deprecate it by
removing the warning.

Test Plan:
built the thrift compiler, then prometheus, and there were no
warnings about "from"

Reviewed By: andrewcox@fb.com

Subscribers: ryandm, alandau

FB internal diff: D1340061