Decouple Header{Client,Server}Channel from Cpp2Channel
Summary:
The common work of the client and server was put in their base
class Cpp2Channel. However, this means that the layering is fixed at
compile time and e.g. on message receive, Cpp2Channel can't decide
whether the client or server needs to handle it.

This diff decouples the client and server channels from Cpp2Channel in a
manner similar to Netty's pipeline.

Cpp2Channel contains handler objects that process incoming and outgoing
messages. The handlers are settable from the outside. Currently, there
are only 2 handlers: encryption and framing.

This is only a refactoring only diff.

Test Plan: fbmake runtests

Reviewed By: davejwatson@fb.com

Subscribers: alandau, bmatheny

FB internal diff: D1357579