Be able to deserialize without throwing an application exception
Summary:
There is no reason to require clients to throw an application exception if one is passed by the server. The client should be able to optionally deserialize and check for an exception in an exception_wrapper. For that, there is now a recv_wrapped_foo() function that is autogenerated.

Until all exceptions are passed to ClientReceiveState() as an exception_wrapper, the most efficient way to handle a response is as follows:

if (crs.isException()) {
handleError();
} else {
auto ew = MyServiceClient::recv_wraped_foo(crs);
if (ew) {
handleError();
}
handleSuccess();
}

Test Plan: unit test

Reviewed By: davejwatson@fb.com

Subscribers: philipp, mig, sdoroshenko, lucian, soren, ruibalp, nli, fugalh, alandau, bmatheny, tudorb

FB internal diff: D1315903