Enhance D's thriftserver.d (CppServer) to expose more functionality
Summary:
For writing effective test fixtures, I wanted to get access to
the event base, the event base manager, and the tsocketaddress.

So I added them.

Test Plan: I wrote a unit test, but I'm not sure if this is the right place for it. Advice welcome.

Reviewed By: davejwatson@fb.com

Subscribers: alandau, bmatheny, yfeldblum

FB internal diff: D1352031