Set KRB5_CONFIG env variable before constructing Krb5Context
Summary: Since the CredentialsCacheManager is constructed globally, it's done before the handshake actually starts. This means the env variable for the config is not yet set. I set it in this diff. Note that I had to make Krb5Context a unique_ptr because I needed to call the setenv method before it's constructed.

Test Plan: Built / ran. Made sure the right config file is used.

Reviewed By: mhorowitz@fb.com

Subscribers: alandau, bmatheny, rodrigos, srenfro, wstefancik

FB internal diff: D1367981