Default connections to sys.<prefix> principals, as opposed to tier names
Summary:
This is a better default overall. We deploy keys to all machines of the form sys.<hostnameprefix>. Before this change, the connections defaulted to the tier's name as opposed to this principal name. This isn't good in most cases.

For custom keys, you're still able to override with the thrift_security_service_tier property.

Test Plan: loadgen

Reviewed By: mhorowitz@fb.com

Subscribers: mcduff, marccelani, mshneer, alandau, hitesh, rodrigos, wstefancik

FB internal diff: D1375675