Add some more useful logs / remove unused ones
Summary:
Rather than saying what I added / removed, I'll just document the different log entries.

Two main ones that are always there:
sasl_ticket_fetch_latency - the time it takes to fetch a ticket. This number is the total and can include either going to the KDC, going to memory, or waiting for disk.
sasl_kdc_latency - if we do end up going to the kdc, this number will just record the time it took. This is not always set, but when it is, it is a good proxy for looking at our kdc performance

Debug ones:
t:FirstRtt,t:SecondRtt,t:ThirdRtt - the time the client spends waiting for each leg of the handshake. These represent network delay / server speed
t:PrepareFirstRequest - total time it took to prepare the initial request. This may include the time it took to fetch tickets, dns resolution, name parsing, waiting on the thread manager, etc.
t:WaitForTgt - time we spend waiting for a valid tgt. This time may include waiting for file parsing or a trip to the KDC.
t:InitBarebonesCcache - time it takes to init a ccache only containing the tgt. There may be contention around this, so this keeps track of that
t:ThreadManagerOverhead - overhead around having a thread manager, and any delays caused by it
t:EvbOverhead - delays caused by busy EVB's. Since the SASL handshake uses the evb queue, and every step of it schedules stuff back on the queue, any delays in the evb could be serious.

Test Plan: loadgen, looked at logs

Reviewed By: acampi@fb.com

Subscribers: akhi, mcduff, marccelani, mshneer, alandau, bmatheny, rodrigos, wstefancik

FB internal diff: D1389032