propagate exceptions from the requestReceived callback
Summary:
If the requestReceived callback throws and is not propagated,
then the caller never gets an event to continue processing.  It will
have to time out, which is slow and not helpful.

Test Plan:
When there's bugs in TCppServer, this was happening all the
time.

Reviewed By: davejwatson@fb.com

Subscribers: alandau, bmatheny

FB internal diff: D1455887

Tasks: 2057547