Change Thrift codegen to work with dmd 2.065
Summary: In 2.065 the D front-end became more trigger-happy about detecting recursive instantiation of templates, so this diff special-cases two more functions. The question remains open whether this is an insufficiency in introspection or a misapplication thereof.

Test Plan: built it, ran it, became a part of it

Reviewed By: davidsp@fb.com, simpkins@fb.com

Subscribers: alandau, bmatheny

FB internal diff: D1443169

Tasks: 4503363