Method to find task expiration time for a request
Summary: The code to compare client timeout with server task expiration time, and use the smaller among the two is hidden in `Cpp2Connection.cpp`. This commit refactors the code to move this to a public method to `ThriftServer.cpp`. This is being done because there are now atleast two places where we would like to find out the task expiration time for a particular request. One is `Cpp2Connection.cpp` (as seen in this diff), and the other is in the moonshot project (please see diff D1453218).

Test Plan: Existing unit tests passed.

Reviewed By: andrewcox@fb.com

Subscribers: alandau, bmatheny, kowshik, shuojie, john

FB internal diff: D1455218

Tasks: 4748883