Fix an ordering bug in ThriftServer's call to preServe.
Summary:
[Thrift] Fix an ordering bug in ThriftServer's call to preServe.

The `preServe` method is what callers use to wait for the ThriftServer to be sufficiently ready. Previously, at the point where `preServe` was called, the server was ready to allow incoming connections. However, it was not ready to terminate the server. In fact, attempting to terminate the server would block indefinitely. After this change, we will be ready to terminate the server immediately after calling `preServe`.

Test Plan: Unit tests.

Reviewed By: andrewcox@fb.com

Subscribers: alandau, bmatheny, dougw

FB internal diff: D1463802

Tasks: 4798695