Added HTTP tunnel thrift clients
Summary: For offnet FNAs, we need a way to communicate to the thrift ports via fwdproxy. We can start with a HTTP tunnel via CONNECT and go from there. This is essentially what the ProxyCommand to ssh into FNAs does.

Test Plan: tested via connecting to fna001.01.fna2914 and the thrift clients are working.

Reviewed By: brianp@fb.com

Subscribers: alandau, bmatheny, shuff, ruibalp

FB internal diff: D1467536

Tasks: 4746093