Use correct protocol (from header) to encode auth messages
Summary:
Before this diff, auth messages (the handshake phase) were
always using compact protocol. If the user specified in the header that
they use binary, then the auth messages had binary in the header but the
payload was still compact.

C++ doesn't care because it always assumes that the auth payload is
compact, but Java cares. So a C++ client that used header/binary
couldn't talk to a Java server (with auth enabled), and vice versa.

This diff changes the C++ client and server to use the correct protocol
(specified in the header) for auth messages (payloads).
It also provides a fallback to be able to talk to old
C++ clients/servers (if using binary and deserialization of an auth
message fails with a protocol error, try deserializing with compact)

Test Plan:
Tested the following
c++ client -> java server
old c++ -> new c++
new c++ -> new c++

explicitly verified that the fallback case is hit and works

Reviewed By: haijunz@fb.com, mhorowitz@fb.com

Subscribers: alandau, bmatheny

FB internal diff: D1486370