Don't log internal exceptions when we process a oneway request
Summary:
If a request returns an empty buffer, assume it's oneway and
don't call req->sendReply().  Since the client doesn't always
correctly indicate oneway-ness, this is the best we can do.

Test Plan:
with the new one way test in cpp_server_test.lpar, verify
that it doesn't log anything (as it did prior to the server fixes).

Reviewed By: haijunz@fb.com

Subscribers: ruibalp, alandau, bmatheny

FB internal diff: D1486724

Tasks: 4861179