Exclude sasl handshake failures from errorTowardsCapping if the policy is set to permitted
Summary:
When the thrift policy is set to permitted, the sasl handshake failures
should not count towards error capping. This will avoid the bug reported in the
task.

Also, for the channel, clear the "suppress fallback to insecure" before trying
to set it.

Test Plan:
fbconfig -r servicerouter && fbmake runtests
Tested manually as per the comments on the task that after the fix, the bug was
not reproduced.

Reviewed By: hitesh@fb.com

Subscribers: alandau, bmatheny, mcduff, hitesh, mshneer, marccelani, mhorowitz

FB internal diff: D1487832

Tasks: 4828416