another fix (old client -> new server)
Summary:
Not sure why it worked before and then broke.
If the server receives a compact message with a binary header, it tries
to deserialize with binary, fails, retries with compact and (the new
thing in this diff) remembers to use compact for the reply.
Previously the reply would have been serialized with the protocol from
the header (which broke old binary clients)

Facebook:
@override-unit-failures
- it thinks hphp_dbg_contbuild_push_blocking is still pending, while
sandcastle says it finished successfully

Test Plan: tests old client -> new server

Reviewed By: haijunz@fb.com

Subscribers: alandau, bmatheny

FB internal diff: D1498610

Tasks: 4914390