limit the number of outstanding secure connects; manage timeouts better
Summary:
Modify the thread mananger to queue connection setups if too
many are pending.  This also adds a callback to indicate when the
connection setup is actually starting, so timeouts can work more
meaningfully when a lot of connections are started at once.  This is a
resubmit of D1479164 which caused perflab errors (see t4925245) and
D1503561 which caused hive canary crashes (see t4947760). It includes
changes to stash a copy of saslThreadManager_, and to hold a mutex
when accessing evb_ to guarantee the channel remains available for the
duration of the call.  It also adds an interface to log an integer
value, so we can track stats better in ODS.

Test Plan:
Test fanout to many servers at once; confirm that all
expected connections succeed.  Run some perflab experiments as
described in t4925245, make sure they succeed.  Hack the client to
drop some messages so timeouts happen; confirm the only additional
errors are these timeouts.

Reviewed By: andrei.bajenov@fb.com

Subscribers: mcduff, hitesh, mshneer, marccelani, alandau, bmatheny, trunkagent

FB internal diff: D1510998

Tasks: 4793859, 4925245