Remove intermediate hop in ThreadManager
Summary:
Maybe this made sense with some old lockless queue, but it doesn't make sense with folly::MPMCQueue + LifoSem anymore.

The tradeoff here is that we are getting rid of an extra thread hop and context switch, but when there are a huge number of contending threads, there is additional overhead to the queue. I don't know of many use cases in practice where we are hammering a queue so hard from so many threads, so I think its safe to say that removing the extra thread is a win we should take.

Test Plan: queue benchmark

Reviewed By: alandau@fb.com

Subscribers: fbcode-common-diffs@, alandau, bmatheny

FB internal diff: D1509568