Fix compiler issue with fuzzing binary fields
Summary:
Binary fields get treated as strings in this set. This allows
them to be treated as binary so the t_type doesn't turn into being the
first one it finds. Feel this solution's not too elegant, but it's the
best I can think of... See P16124959 for an example of the problem and
solution

Test Plan: Patch D1532586 - saw error before, now it compiles cleanly

Reviewed By: watashi@fb.com

Subscribers: alandau

FB internal diff: D1532593

Tasks: 4773565