Fix getLastUsedHost in Java interface when using header transport
Summary:
getLastUsedHost gave a casting error when we turned on header protocol.  With this change, we'll be calling getLastUsedHost on the underlying ServiceRouterTransport, not the HeaderTransport wrapping it.

Tried to follow the same basic structure as D1493801

Test Plan: Testing the particular change w/ a unit test (passed), other tests are failing but I think they're flaky...

Reviewed By: alandau@fb.com

Subscribers: mcduff, marccelani, hitesh, mshneer, alandau, bmatheny

FB internal diff: D1535778

Tasks: 5051850, 4988429