Add support for persistent_headers in Java
Summary:
In Serf, we need some headers to be sent with every request to function properly.
e.g., we send your used id in a header to know who is doing what, and check for permissions.

We need these headers to be on *every* request sent by our clients, therefore we must use `setPersistentHeader` instead of `setHeader`.

Unfortunately, persistentHeaders are not implemented in Java. This diff implements it.

Gotcha:
* writerHeaders are now erased after each requests. replace `setHeaders` with `setPersistentHeaders` if that's not what you want

Test Plan:
I'm testing this with my diffs for THeader support in SerfUtil: D1510765 and D1509281

Specifically, I've added tests that use the same client twice.

Reviewed By: qinteng@fb.com, alandau@fb.com

Subscribers: gangli, ruibalp, alandau, bmatheny, sajalj

FB internal diff: D1546292

Tasks: 5125024