FunctionSendCallback.
Summary:
[Thrift] FunctionSendCallback.

A callback object for callers interested in knowing when a message is sent, or when a send error is encountered. In particular, this is useful for callers sending oneway rpc messages. This might even be named `OnewayCallback`.

This is a convenience class allowing callers to pass a lambda expression instead of having to define an ad-hoc callback class and construct an instance.

Test Plan: Unit tests in `thrift/lib/cpp2/test/RequestChannelTest.cpp`.

Reviewed By: davejwatson@fb.com

Subscribers: trunkagent, alandau, bmatheny, dougw

FB internal diff: D1571459