Revert a special-case race fix which is causing failures with an EVB change.
Summary:
This reverts a fix (D1241245) to a race condition which was
happening in July.  This race is probably the same one which let to me
writing D1455885, and the two fixes are not compatible.  As the more
recent change is more generic, this diff reverts the older fix.  The
fix in the older diff still passes.

More specifically, the problem is that if the non-sync call fails (for
example, if the server is down), before the loop is invoked, the test
for _returnState.exception() before the loop will pass, and the state
set by terminateLoopSoon() in the EventBase will remain set.  On the
next call, if there is no exception before the loop, the loop will be
a noop, neither the error nor return fields in the state will be set,
and the assert() will happen.  The fix is to run the event loop
always, so if terminateLoopSoon() was called, the state it left behind
will be reset.

Without this change, the simplest way to repro this is to call a
method twice, so the client is bad the second time through with the
same evb.  This diff also adds a test to ThriftServerTest which does
that.

Test Plan:
_build/dbg/thrift/lib/cpp2/test/ThriftServerTest, and the
test described in #5217793

Reviewed By: davejwatson@fb.com

Subscribers: bmatheny, bsimmers, trunkagent, ruibalp, alandau, marccelani

FB internal diff: D1572519

Tasks: 5217793

Blame Revision: D1241245