move shutdownsocketset out of asncoscket
Summary:
Moving async socket to folly, this is the only dep not in folly.  Not convinced this *needs* to be in async socket, as opposed to just the thrift server - it doesn't appear to be used anywhere except in unicorn thrift server.

It ends up being less code this way too.

Test Plan:
fbconfig -r thrift; fbmake runtests

The unittest for this don't look amazing, maybe I should write another

Reviewed By: alandau@fb.com

Subscribers: zeus-diffs@, tw-eng@, benj, scuba-diffs@, trunkagent, doug, rodrigos, alandau, bmatheny, acampi, dcsommer

FB internal diff: D1583629