rm clang as a host-specific option
Summary:
This precludes us from building folly -r with clang, since it includes
a binary that is both a tool and a regular binary. This is a general problem,
and the clang fix was a hack. It is possible that the clang hack fix is still
necessary, but it is about a year old, so we should verify that assertion.

Test Plan:
fbconfig tools/build/test/tests && fbmake runtests
phabricator unit tests

Reviewed By: robbert@fb.com

Subscribers: alandau, trunkagent, simpkins, agallagher, ldbrandy, sdwilsh, mathieubaudet, mpawlowski

FB internal diff: D1601468

Tasks: 5324999